+++ spec/municipitaly_spec.rb
@@ -2,6 +2,6 @@
 
 RSpec.describe Municipitaly do
   it 'has a version number' do
-    expect(Municipitaly::VERSION).to be '0.4.0'
+    expect(Municipitaly::VERSION).to be '0.5.0'
   end
 end