+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2024-06-20 (forked by yupswing)
+
+### Added
+
+- Fixed search provinces or municipalities with diacritics
+- Updated database 27-05-2024 (population data as 2021)\
+- Municipalities `latitude`, `longitude` and `area` (km2)
+
 ## [0.4.0] - 2020-11-18
 
 ### Added
@@ -23,16 +31,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - ISO 3166-2 code in province object `#iso3166_2` or alias `#iso3166`
 
-
 ## [0.1.0] - 2020-02-13
 
 ### Changed
 
 - Update municipalities data according to latest ISTAT changes (jan 2020)
 
-
 ## [0.0.3] - 2020-02-10
 
 ### Added
 
-- The optional parameter *greedy* in `Municipitaly::Search.municipalities_from_name`
+- The optional parameter _greedy_ in `Municipitaly::Search.municipalities_from_name`